# backend/internal/httpapi/openapi.yaml
openapi: 3.0.3
info:
  title: KubeLens API
  version: v0.2.0
  description: Core API contract for KubeLens dashboard and automation clients.
servers:
  - url: /api
paths:
...
  /pods:
    get:
      operationId: listPods
      summary: List pods across namespaces
      responses:
        "200":
          description: Pod list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PodSummary"
    post:
      operationId: createPod
      summary: Create pod
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PodCreateRequest"
      responses:
        "200":
          description: Action result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /pods/{namespace}/{name}:
    get:
      operationId: getPodDetail
      summary: Get pod detail
      parameters:
        - in: path
          name: namespace
          required: true
          schema:
            type: string
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pod detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PodDetail"
    delete:
      operationId: deletePod
      summary: Delete pod
      parameters:
        - in: path
          name: namespace
          required: true
          schema:
            type: string
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Action result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
...
  /pods/{namespace}/{name}/logs/stream:
    get:
      operationId: streamPodLogs
      summary: Stream pod logs
      parameters:
        - in: path
          name: namespace
          required: true
          schema:
            type: string
        - in: path
          name: name
          required: true
          schema:
            type: string
        - in: query
          name: lines
          required: false
          schema:
            type: integer
            minimum: 1
            default: 50
        - in: query
          name: container
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Streaming log output
          content:
            text/plain:
              schema:
                type: string
...
  /pods/{namespace}/{name}/restart:
    post:
      operationId: restartPod
      summary: Restart pod
      parameters:
        - in: path
          name: namespace
          required: true
          schema:
            type: string
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Action result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
...
  /nodes/{name}/cordon:
    post:
      operationId: cordonNode
      summary: Cordon node
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Action result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
  /resources/{kind}:
    get:
      operationId: listResources
      summary: List resources by kind
      parameters:
        - in: path
          name: kind
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resource list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceList"
  /resources/{kind}/{namespace}/{name}/yaml:
    get:
      operationId: getResourceYAML
      summary: Fetch resource YAML
      parameters:
        - in: path
          name: kind
          required: true
          schema:
            type: string
        - in: path
          name: namespace
          required: true
          schema:
            type: string
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resource manifest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceManifest"
    put:
      operationId: putResourceYAML
      summary: Apply resource YAML
      parameters:
        - in: path
          name: kind
          required: true
          schema:
            type: string
        - in: path
          name: namespace
          required: true
          schema:
            type: string
        - in: path
          name: name
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceManifest"
      responses:
        "200":
          description: Action result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    PodSummary:
      type: object
      required: [id, name, namespace, status, cpu, memory, age, restarts]
      properties:
        id:
          type: string
        name:
          type: string
        namespace:
          type: string
        status:
          type: string
        cpu:
          type: string
        memory:
          type: string
        age:
          type: string
        restarts:
          type: integer
    PodDetail:
      allOf:
        - $ref: "#/components/schemas/PodSummary"
        - type: object
          properties:
            nodeName:
              type: string
            hostIP:
              type: string
            podIP:
              type: string
            containers:
              type: array
              items:
                $ref: "#/components/schemas/ContainerSpec"
            volumes:
              type: array
              items:
                $ref: "#/components/schemas/NamedVolume"
    ContainerSpec:
      type: object
      required: [name]
      properties:
        name:
          type: string
        image:
          type: string
        env:
          type: array
          items:
            $ref: "#/components/schemas/EnvVar"
        volumeMounts:
          type: array
          items:
            $ref: "#/components/schemas/VolumeMount"
        resources:
          $ref: "#/components/schemas/ContainerResources"
    EnvVar:
      type: object
      required: [name]
      properties:
        name:
          type: string
        value:
          type: string
    VolumeMount:
      type: object
      properties:
        name:
          type: string
        mountPath:
          type: string
    ContainerResources:
      type: object
      properties:
        requests:
          $ref: "#/components/schemas/ResourcePairs"
        limits:
          $ref: "#/components/schemas/ResourcePairs"
    ResourcePairs:
      type: object
      properties:
        cpu:
          type: string
        memory:
          type: string
    NamedVolume:
      type: object
      properties:
        name:
          type: string
...
    ResourceRecord:
      type: object
      required: [id, name, status, age]
      properties:
        id:
          type: string
        name:
          type: string
        namespace:
          type: string
        status:
          type: string
        age:
          type: string
        summary:
          type: string
    ResourceList:
      type: object
      required: [kind, items]
      properties:
        kind:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ResourceRecord"
    ResourceManifest:
      type: object
      required: [yaml]
      properties:
        yaml:
          type: string
    PodCreateRequest:
      type: object
      required: [namespace, name, image]
      properties:
        namespace:
          type: string
        name:
          type: string
        image:
          type: string
...
    ActionResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
...